+++ Services/CourseSystem.Services.Data/DecksService.cs
@@ -7,16 +7,19 @@
     using CourseSystem.Data.Common.Repositories;
     using CourseSystem.Data.Models;
     using CourseSystem.Services.Mapping;
+    using Microsoft.AspNetCore.Identity;
 
     public class DecksService : IDecksService
     {
         private readonly IDeletableEntityRepository<Deck> decksRepository;
         private readonly IDeletableEntityRepository<Card> cardsRepository;
+        private readonly UserManager<ApplicationUser> userManager;
 
-        public DecksService(IDeletableEntityRepository<Deck> decksRepository, IDeletableEntityRepository<Card> cardsRepository)
+        public DecksService(IDeletableEntityRepository<Deck> decksRepository, IDeletableEntityRepository<Card> cardsRepository, UserManager<ApplicationUser> userManager)
         {
             this.decksRepository = decksRepository;
             this.cardsRepository = cardsRepository;
+            this.userManager = userManager;
         }
 
         public async Task CreateCard(string frontSide, string backSide, string deckId)
@@ -55,18 +58,25 @@
             return deck;
         }
 
-        public async Task DeleteCard(string id)
+        public async Task DeleteCard(string id, string userId)
         {
             var card = this.cardsRepository.All().FirstOrDefault(x => x.Id == id);
-            this.cardsRepository.Delete(card);
-            await this.cardsRepository.SaveChangesAsync();
+            var deck = this.decksRepository.All().FirstOrDefault(x => x.Id == card.DeckId);
+            if (deck.UserId == userId)
+            {
+                this.cardsRepository.Delete(card);
+                await this.cardsRepository.SaveChangesAsync();
+            }
         }
 
-        public async Task DeleteDeck(string id)
+        public async Task DeleteDeck(string id, string userId)
         {
             var deck = this.decksRepository.All().FirstOrDefault(x => x.Id == id);
-            this.decksRepository.Delete(deck);
-            await this.decksRepository.SaveChangesAsync();
+            if (deck.UserId == userId)
+            {
+                this.decksRepository.Delete(deck);
+                await this.decksRepository.SaveChangesAsync();
+            }
         }
 
         public T GetCard<T>(string cardId)
@@ -102,16 +112,20 @@
             return decks;
         }
 
-        public async Task UpdateCard(string frontSide, string backSide, string id)
+        public async Task UpdateCard(string frontSide, string backSide, string id, string userId)
         {
             var card = this.cardsRepository.All()
                 .FirstOrDefault(x => x.Id == id);
 
-            card.FrontSide = frontSide;
-            card.BackSide = backSide;
+            var deck = this.decksRepository.All().FirstOrDefault(x => x.Id == card.DeckId);
+            if (deck.UserId == userId)
+            {
+                card.FrontSide = frontSide;
+                card.BackSide = backSide;
 
-            this.cardsRepository.Update(card);
-            await this.cardsRepository.SaveChangesAsync();
+                this.cardsRepository.Update(card);
+                await this.cardsRepository.SaveChangesAsync();
+            }
         }
     }
 }
